Refactor ModuleManager and VanillaPlus to centralize message handling and improve logging format
