+++ src/main/java/org/xodium/vanillaplus/ModuleManager.java
@@ -1,12 +1,12 @@
 package org.xodium.vanillaplus;
 
-import org.xodium.vanillaplus.interfaces.MSG;
+import org.xodium.vanillaplus.VanillaPlus.MSG;
 import org.xodium.vanillaplus.interfaces.Modular;
 import org.xodium.vanillaplus.modules.DoorsModule;
 
 import java.util.List;
 
-public class ModuleManager {
+public class ModuleManager implements MSG {
     private final VanillaPlus vp = VanillaPlus.getInstance();
 
     {
@@ -16,7 +16,7 @@ public class ModuleManager {
                 .filter(Modular::isEnabled)
                 .forEach(module -> {
                     vp.getServer().getPluginManager().registerEvents(module, vp);
-                    vp.getLogger().info(MSG.MODULE_LOADED + module.getClass().getSimpleName());
+                    vp.getLogger().info(PREFIX + "Loaded: " + module.getClass().getSimpleName());
                 });
     }
 }
