+++ src/main/java/org/xodium/vanillaplus/VanillaPlus.java
@@ -9,7 +9,12 @@ public class VanillaPlus extends JavaPlugin {
   private static final String[] V = { "1.21.1", "1.21.3", "1.21.4" };
   private static final String[] PAPER = { "Paper" };
   private static final String IS_PAPER_MSG = "This plugin is not compatible with non-Paper servers.";
-  private static final String IS_SUPPORTED_VERSION_MSG = "This plugin requires Paper version: " + String.join(", ", V);
+  private static final String IS_SUPPORTED_VERSION_MSG = "This plugin requires Paper version(s): "
+      + String.join(", ", V);
+
+  public interface MSG {
+    String PREFIX = "<gold>[<dark_aqua>VanillaPlus<gold>] <reset>";
+  }
 
   public static VanillaPlus getInstance() {
     return getPlugin(VanillaPlus.class);
