--- src/main/java/org/xodium/vanillaplus/interfaces/MSG.java
@@ -1,6 +1,0 @@
-package org.xodium.vanillaplus.interfaces;
-
-public interface MSG {
-    String PREFIX = "<gold>[<dark_aqua>VanillaPlus<gold>] <reset>";
-    String MODULE_LOADED = "Loaded: ";
-}
